Create form to use for forecast lookup and page to display it on
This creates `ForecastController#index` which renders the form for
looking up the forecast by zip code. We set this to also be the root
path ("/").

It is not a normal form because we want to go to a specific URL with the
zip code in the path and not submit the data as either a GET parameter
or in the body of a POST request. This is achieved through a stimulus
controller that allows for replacing a bit of a URL with the value in
the text field. (This should be generic enough that we could reuse it in
other contexts.)

I did add a system test that tests the entire flow. It visits a page
with the form, puts in a zip code, and then clicks the button. The test
verifies that the new page with the current conditions and forecast
comes up.